Show correct price with discount
