+++ src/components/checkout/CartOverview/script.js
@@ -9,6 +9,7 @@ import LineItemInfo from '../../common/cartlike/LineItemInfo/index.vue';
 import CART_FRAGMENT from '../../Cart.gql';
 import MONEY_FRAGMENT from '../../Money.gql';
 import ADDRESS_FRAGMENT from '../../Address.gql';
+import { totalPrice } from '../../common/shared';
 
 export default {
   components: {
@@ -22,6 +23,9 @@ export default {
   data: () => ({
     me: null,
   }),
+  methods: {
+    totalPrice,
+  },
   apollo: {
     me: {
       query: gql`
