+++ src/components/common/cartlike/CartLikeContentDetail/script.js
@@ -2,6 +2,7 @@ import LineItemInfo from '../LineItemInfo/index.vue';
 import BasePrice from '../../BasePrice/index.vue';
 import LineItemQuantityForm from '../../../cartdetail/LineItemQuantityForm/index.vue';
 import LineItemDeleteForm from '../../../cartdetail/LineItemDeleteForm/index.vue';
+import { totalPrice } from '../../shared';
 
 export default {
   components: {
@@ -21,17 +22,6 @@ export default {
     },
   },
   methods: {
-    totalPrice(lineItem) {
-      const { centAmount: unitCentAmount, ...unitPrice } = lineItem.price.discounted?.value || lineItem.price.value;
-      const originalPrice = {
-        ...unitPrice,
-        centAmount: unitCentAmount * lineItem.quantity,
-      };
-      const price = { value: { ...originalPrice } };
-      if (originalPrice.centAmount !== lineItem.totalPrice.centAmount) {
-        price.discounted = { value: { ...lineItem.totalPrice } };
-      }
-      return price;
-    },
+    totalPrice,
   },
 };
